FAQ for simple / common errors
Some proposed FAQ questions and answers based on my initial experience working with the pipeline of places I made errors. Both are probably unnecessary but may help reduce the number of simple questions.

It may be a good idea to generate an alternative place for FAQ's as the documentation on this page increases.